streamlit_resume_matcher/requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
plotly==5.17.0
pandas==2.1.4